+++ packages/components/date-input/__tests__/time-input.test.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable jsx-a11y/no-autofocus */
 import * as React from "react";
-import {act, fireEvent, render} from "@testing-library/react";
+import {fireEvent, render} from "@testing-library/react";
 import {Time, ZonedDateTime} from "@internationalized/date";
 import {TimeValue} from "@react-types/datepicker";
 import {pointerMap, triggerPress} from "@nextui-org/test-utils";
@@ -180,6 +180,32 @@ describe("TimeInput", () => {
         }
       }
     });
+
+    it("should support error message (with isInvalid)", function () {
+      const {getAllByRole, getByRole} = render(
+        <TimeInput isInvalid errorMessage="Error message" label="Time" />,
+      );
+
+      const group = getByRole("group");
+
+      expect(group).toHaveAttribute("aria-describedby");
+
+      if (group) {
+        const descById = group.getAttribute("aria-describedby");
+        const description = descById && document.getElementById(descById);
+
+        expect(description).toHaveTextContent("Error message");
+
+        const segments = getAllByRole("spinbutton");
+
+        for (const segment of segments) {
+          expect(segment).toHaveAttribute(
+            "aria-describedby",
+            group.getAttribute("aria-describedby"),
+          );
+        }
+      }
+    });
   });
 
   describe("Events", function () {
@@ -211,17 +237,13 @@ describe("TimeInput", () => {
       expect(onBlurSpy).not.toHaveBeenCalled();
       expect(onFocusChangeSpy).not.toHaveBeenCalled();
       expect(onFocusSpy).not.toHaveBeenCalled();
-      await act(async () => {
-        await user.tab();
-      });
+      await user.tab();
       expect(segments[0]).toHaveFocus();
 
       expect(onBlurSpy).not.toHaveBeenCalled();
       expect(onFocusChangeSpy).toHaveBeenCalledTimes(1);
       expect(onFocusSpy).toHaveBeenCalledTimes(1);
-      await act(async () => {
-        await user.tab();
-      });
+      await user.tab();
       expect(segments[1]).toHaveFocus();
       expect(onBlurSpy).not.toHaveBeenCalled();
       expect(onFocusChangeSpy).toHaveBeenCalledTimes(1);
@@ -242,22 +264,14 @@ describe("TimeInput", () => {
       expect(onBlurSpy).not.toHaveBeenCalled();
       expect(onFocusChangeSpy).not.toHaveBeenCalled();
       expect(onFocusSpy).not.toHaveBeenCalled();
-      await act(async () => {
-        await user.tab();
-      });
+      await user.tab();
       expect(segments[0]).toHaveFocus();
-      await act(async () => {
-        await user.tab();
-      });
+      await user.tab();
       expect(segments[1]).toHaveFocus();
-      await act(async () => {
-        await user.tab();
-      });
+      await user.tab();
       expect(segments[2]).toHaveFocus();
       expect(onBlurSpy).toHaveBeenCalledTimes(0);
-      await act(async () => {
-        await user.tab();
-      });
+      await user.tab();
       expect(onBlurSpy).toHaveBeenCalledTimes(1);
       expect(onFocusChangeSpy).toHaveBeenCalledTimes(2);
       expect(onFocusSpy).toHaveBeenCalledTimes(1);
@@ -272,10 +286,7 @@ describe("TimeInput", () => {
       expect(onKeyDownSpy).not.toHaveBeenCalled();
       expect(onKeyUpSpy).not.toHaveBeenCalled();
 
-      await act(() => {
-        user.tab();
-      });
-
+      await user.tab();
       expect(segments[0]).toHaveFocus();
       expect(onKeyDownSpy).not.toHaveBeenCalled();
       expect(onKeyUpSpy).toHaveBeenCalledTimes(1);
@@ -352,6 +363,56 @@ describe("TimeInput", () => {
 
       expect(getDescription()).toBe("Selected Time: 8:30 AM");
       expect(input).toHaveValue("08:30:00");
+    });
+  });
+
+  describe(`Validation (validationBehavior="aria")`, () => {
+    it("should display errorMessage when timeValue is less than the minimum (controlled)", () => {
+      render(<TimeInput label="Time" minValue={new Time(9)} value={new Time(8)} />);
+
+      expect(document.querySelector("[data-slot=error-message]")).toBeVisible();
+    });
+
+    it("should display errorMessage when timeValue is less than the minimum (uncontrolled)", async () => {
+      const {getAllByRole} = render(
+        <TimeInput defaultValue={new Time(9)} label="Time" minValue={new Time(9)} name="time" />,
+      );
+
+      const input = document.querySelector("input[name=time]");
+      const segments = getAllByRole("spinbutton");
+
+      await user.tab();
+      expect(input).toHaveValue("09:00:00");
+      expect(segments[0]).toHaveFocus();
+      expect(document.querySelector("[data-slot=error-message]")).toBeNull();
+
+      await user.keyboard("[ArrowDown]");
+      expect(input).toHaveValue("08:00:00");
+      expect(document.querySelector("[data-slot=error-message]")).toBeVisible();
+    });
+
+    it("should display errorMessage when timeValue is greater than the maximum (controlled)", () => {
+      render(<TimeInput label="Time" maxValue={new Time(17)} value={new Time(18)} />);
+
+      expect(document.querySelector("[data-slot=error-message]")).toBeVisible();
+    });
+
+    it("should display errorMessage when timeValue is greater than the maximum (uncontrolled)", async () => {
+      const {getAllByRole} = render(
+        <TimeInput defaultValue={new Time(17)} label="Time" maxValue={new Time(17)} name="time" />,
+      );
+
+      const input = document.querySelector("input[name=time]");
+      const segments = getAllByRole("spinbutton");
+
+      await user.tab();
+      expect(input).toHaveValue("17:00:00");
+      expect(segments[0]).toHaveFocus();
+      expect(document.querySelector("[data-slot=error-message]")).toBeNull();
+
+      await user.keyboard("[ArrowUp]");
+      expect(input).toHaveValue("18:00:00");
+      expect(document.querySelector("[data-slot=error-message]")).toBeVisible();
     });
   });
 });