+++ packages/components/date-input/src/use-date-input.ts
@@ -1,22 +1,22 @@
 import type {DateInputVariantProps, DateInputSlots, SlotsToClasses} from "@nextui-org/theme";
 import type {AriaDateFieldProps} from "@react-types/datepicker";
 import type {SupportedCalendars} from "@nextui-org/system";
-import type {DateValue, Calendar} from "@internationalized/date";
+import type {DateValue} from "@react-types/datepicker";
+import type {Calendar} from "@internationalized/date";
 import type {ReactRef} from "@nextui-org/react-utils";
 import type {DOMAttributes, GroupDOMAttributes} from "@react-types/shared";
 import type {DateInputGroupProps} from "./date-input-group";
 
 import {useLocale} from "@react-aria/i18n";
-import {CalendarDate} from "@internationalized/date";
+import {createCalendar, CalendarDate, DateFormatter} from "@internationalized/date";
 import {mergeProps} from "@react-aria/utils";
 import {PropGetter, useProviderContext} from "@nextui-org/system";
 import {HTMLNextUIProps, mapPropsVariants} from "@nextui-org/system";
 import {useDOMRef} from "@nextui-org/react-utils";
 import {useDateField as useAriaDateField} from "@react-aria/datepicker";
 import {useDateFieldState} from "@react-stately/datepicker";
-import {createCalendar} from "@internationalized/date";
-import {objectToDeps, clsx, dataAttr} from "@nextui-org/shared-utils";
-import {dateInput} from "@nextui-org/theme";
+import {objectToDeps, clsx, dataAttr, getGregorianYearOffset} from "@nextui-org/shared-utils";
+import {dateInput, cn} from "@nextui-org/theme";
 import {useMemo} from "react";
 
 type NextUIBaseProps<T extends DateValue> = Omit<
@@ -35,6 +35,8 @@ interface Props<T extends DateValue> extends NextUIBaseProps<T> {
   labelProps?: DOMAttributes;
   /** Props for the date field. */
   fieldProps?: DOMAttributes;
+  /** Props for the inner wrapper. */
+  innerWrapperProps?: DOMAttributes;
   /** Props for the description element, if any. */
   descriptionProps?: DOMAttributes;
   /** Props for the error message element, if any. */
@@ -116,6 +118,15 @@ export function useDateInput<T extends DateValue>(originalProps: UseDateInputPro
 
   const [props, variantProps] = mapPropsVariants(originalProps, dateInput.variantKeys);
 
+  const {locale} = useLocale();
+
+  const calendarProp = createCalendar(new DateFormatter(locale).resolvedOptions().calendar);
+
+  // by default, we are using gregorian calendar with possible years in [1900, 2099]
+  // however, some locales such as `th-TH-u-ca-buddhist` using different calendar making the years out of bound
+  // hence, add the corresponding offset to make sure the year is within the bound
+  const gregorianYearOffset = getGregorianYearOffset(calendarProp.identifier);
+
   const {
     ref,
     as,
@@ -130,12 +141,15 @@ export function useDateInput<T extends DateValue>(originalProps: UseDateInputPro
     groupProps = {},
     labelProps: labelPropsProp,
     fieldProps: fieldPropsProp,
+    innerWrapperProps: innerWrapperPropsProp,
     errorMessageProps: errorMessagePropsProp,
     descriptionProps: descriptionPropsProp,
     validationBehavior = globalContext?.validationBehavior ?? "aria",
     shouldForceLeadingZeros = true,
-    minValue = globalContext?.defaultDates?.minDate ?? new CalendarDate(1900, 1, 1),
-    maxValue = globalContext?.defaultDates?.maxDate ?? new CalendarDate(2099, 12, 31),
+    minValue = globalContext?.defaultDates?.minDate ??
+      new CalendarDate(calendarProp, 1900 + gregorianYearOffset, 1, 1),
+    maxValue = globalContext?.defaultDates?.maxDate ??
+      new CalendarDate(calendarProp, 2099 + gregorianYearOffset, 12, 31),
     createCalendar: createCalendarProp = globalContext?.createCalendar ?? null,
     isInvalid: isInvalidProp = validationState ? validationState === "invalid" : false,
     errorMessage,
@@ -145,8 +159,6 @@ export function useDateInput<T extends DateValue>(originalProps: UseDateInputPro
   const inputRef = useDOMRef(inputRefProp);
 
   const disableAnimation = originalProps.disableAnimation ?? globalContext?.disableAnimation;
-
-  const {locale} = useLocale();
 
   const state = useDateFieldState({
     ...originalProps,
@@ -243,11 +255,13 @@ export function useDateInput<T extends DateValue>(originalProps: UseDateInputPro
   };
 
   const getInnerWrapperProps: PropGetter = (props) => {
+    const innerWrapperProps = mergeProps(innerWrapperPropsProp, props);
+
     return {
-      ...props,
+      ...innerWrapperProps,
       "data-slot": "inner-wrapper",
       className: slots.innerWrapper({
-        class: classNames?.innerWrapper,
+        class: cn(classNames?.innerWrapper, innerWrapperProps?.className),
       }),
     };
   };