+++ packages/components/date-picker/src/use-date-range-picker.ts
@@ -20,7 +20,7 @@ import {useDateRangePickerState} from "@react-stately/datepicker";
 import {useDateRangePicker as useAriaDateRangePicker} from "@react-aria/datepicker";
 import {clsx, dataAttr, objectToDeps} from "@nextui-org/shared-utils";
 import {mergeProps} from "@react-aria/utils";
-import {dateRangePicker, dateInput} from "@nextui-org/theme";
+import {dateRangePicker, dateInput, cn} from "@nextui-org/theme";
 import {ariaShouldCloseOnInteractOutside} from "@nextui-org/aria-utils";
 
 import {useDatePickerBase} from "./use-date-picker-base";
@@ -215,10 +215,11 @@ export function useDateRangePicker<T extends DateValue>({
             props.className,
           ),
         }),
+        shouldCloseOnInteractOutside: popoverProps?.shouldCloseOnInteractOutside
+          ? popoverProps.shouldCloseOnInteractOutside
+          : (element: Element) =>
+              ariaShouldCloseOnInteractOutside(element, popoverTriggerRef, state),
       },
-      shouldCloseOnInteractOutside: popoverProps?.shouldCloseOnInteractOutside
-        ? popoverProps.shouldCloseOnInteractOutside
-        : (element: Element) => ariaShouldCloseOnInteractOutside(element, domRef, state),
     } as PopoverProps;
   };
 
@@ -227,8 +228,11 @@ export function useDateRangePicker<T extends DateValue>({
       ...ariaCalendarProps,
       ...calendarProps,
       classNames: {
-        base: slots.calendar({class: classNames?.calendar}),
-        content: slots.calendarContent({class: classNames?.calendarContent}),
+        ...calendarProps.classNames,
+        base: slots.calendar({class: cn(calendarProps?.classNames?.base, classNames?.calendar)}),
+        content: slots.calendarContent({
+          class: cn(calendarProps?.classNames?.content, classNames?.calendarContent),
+        }),
       },
     } as RangeCalendarProps;
   };
@@ -391,7 +395,6 @@ export function useDateRangePicker<T extends DateValue>({
       endContent,
       errorMessage,
       isInvalid,
-      startContent,
       validationDetails,
       validationErrors,
       shouldLabelBeOutside,
@@ -418,6 +421,7 @@ export function useDateRangePicker<T extends DateValue>({
     label: originalProps.label,
     slots,
     classNames,
+    startContent,
     endContent,
     selectorIcon,
     showTimeField,