+++ packages/core/theme/src/components/date-picker.ts
@@ -10,11 +10,11 @@ import {tv} from "../utils/tv";
 const datePicker = tv({
   slots: {
     base: "group w-full",
-    selectorButton: "-mr-2 text-inherit",
+    selectorButton: "-mx-2 text-inherit",
     selectorIcon: "text-lg text-inherit pointer-events-none flex-shrink-0",
     popoverContent: "p-0 w-full",
-    calendar: "w-[var(--calendar-width)] shadow-none",
-    calendarContent: "w-[var(--calendar-width)]",
+    calendar: "w-[calc(var(--visible-months)_*_var(--calendar-width))] shadow-none",
+    calendarContent: "w-[calc(var(--visible-months)_*_var(--calendar-width))]",
     timeInputLabel: "font-medium",
     timeInput: "px-5 pb-4 flex-wrap gap-x-6",
   },